chore(docker): dockering tests to isolate dependencies
When using docker, it is no longer necessary to follow the step-by-step instructions on how to run the backend and frontend before running the API tests and guarantees that the application and tests will always run with the same dependencies regardless of which machine they are run on.
The fewer points of failure in the infrastructure, the easier it is to maintain the project over time.